adding folders for ui, adapters, dto and helpers
