+++ app/src/main/AndroidManifest.xml
@@ -10,14 +10,14 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
-        <activity android:name=".MainSearchActivity">
+        <activity android:name=".ui.mainsearchactivity.MainSearchActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
 
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <activity android:name=".SearchResultActivity"></activity>
+        <activity android:name=".ui.searchresultactivity.SearchResultActivity"></activity>
     </application>
 
 </manifest>
