+++ app/src/main/java/com/bimobject/themayproject/dto/Brand.java
@@ -1,25 +1,22 @@
-package com.bimobject.themayproject;
+package com.bimobject.themayproject.dto;
 
-public class Product{
+public class Brand {
     private String id;
     private String name;
-    private String status;
     private String permalink;
+    private String visibility;
     private String imageUrl;
 
-    private Brand brand;
-
-    public Product() {
+    public Brand() {
     }
 
-    public Product(String id, String name, String status, String permalink, String imageUrl) {
+    public Brand(String id, String name, String permalink, String visibility, String imageUrl) {
         this.id = id;
         this.name = name;
-        this.status = status;
         this.permalink = permalink;
+        this.visibility = visibility;
         this.imageUrl = imageUrl;
     }
-
 
     public String getId() {
         return id;
@@ -37,14 +34,6 @@ public class Product{
         this.name = name;
     }
 
-    public String getStatus() {
-        return status;
-    }
-
-    public void setStatus(String status) {
-        this.status = status;
-    }
-
     public String getPermalink() {
         return permalink;
     }
@@ -53,15 +42,19 @@ public class Product{
         this.permalink = permalink;
     }
 
+    public String getVisibility() {
+        return visibility;
+    }
+
+    public void setVisibility(String visibility) {
+        this.visibility = visibility;
+    }
+
     public String getImageUrl() {
         return imageUrl;
     }
 
     public void setImageUrl(String imageUrl) {
         this.imageUrl = imageUrl;
-    }
-
-    public Brand getBrand() {
-        return brand;
     }
 }
