+++ app/src/main/java/com/bimobject/themayproject/helpers/ProductDeserializer.java
@@ -1,5 +1,6 @@
-package com.bimobject.themayproject;
+package com.bimobject.themayproject.helpers;
 
+import com.bimobject.themayproject.dto.Product;
 import com.google.gson.JsonDeserializationContext;
 import com.google.gson.JsonDeserializer;
 import com.google.gson.JsonElement;
