+++ app/src/main/java/com/bimobject/themayproject/helpers/RequestService.java
@@ -1,5 +1,6 @@
-package com.bimobject.themayproject;
+package com.bimobject.themayproject.helpers;
 
+import com.bimobject.themayproject.dto.Product;
 import com.google.gson.Gson;
 import com.google.gson.GsonBuilder;
 import com.google.gson.reflect.TypeToken;
