+++ app/src/main/res/layout/activity_main_search.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.bimobject.themayproject.MainSearchActivity"
+    tools:context="com.bimobject.themayproject.ui.mainsearchactivity.MainSearchActivity"
     tools:layout_editor_absoluteY="81dp">
 
     <EditText
